edit podspec

+++ Demo/相册Demo/Example/Example0ViewController.m
@@ -86,7 +86,7 @@
 }
 
 #pragma mark - <UITableViewDataSource>
-- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
+- (long) tableView:(UITableView *)tableView numberOfRowsInSection:(long)section{
     return self.assets.count;
 }
