+++ Demo/相册Demo/Example/Example1ViewController.m
@@ -82,7 +82,7 @@
 }
 
 #pragma mark - <UITableViewDataSource>
-- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
+- (long) tableView:(UITableView *)tableView numberOfRowsInSection:(long)section{
     return self.photos.count;
 }
 
@@ -130,11 +130,11 @@
 }
 
 #pragma mark - <ZLPhotoPickerBrowserViewControllerDataSource>
-//- (NSInteger)numberOfSectionInPhotosInPickerBrowser:(ZLPhotoPickerBrowserViewController *)pickerBrowser{
+//- (long)numberOfSectionInPhotosInPickerBrowser:(ZLPhotoPickerBrowserViewController *)pickerBrowser{
 //    return 1;
 //}
 //
-//- (NSInteger)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
+//- (long)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
 //    return self.assets.count;
 //}
 //
