+++ ZLPhotoLib/Classes/ZLPhotoPicker/ViewControllers/ZLPhotoPickerAssetsViewController.m
@@ -121,7 +121,7 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
     self.collectionView.lastDataArray = nil;
     self.collectionView.isRecoderSelectPicker = YES;
     self.collectionView.selectAssets = self.selectAssets;
-    NSInteger count = self.selectAssets.count;
+    long count = self.selectAssets.count;
     self.makeView.hidden = !count;
     self.makeView.text = [NSString stringWithFormat:@"%ld",(long)count];
     self.doneBtn.enabled = (count > 0);
@@ -266,7 +266,7 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
         [self.toolBarThumbCollectionView reloadData];
         [self.takePhotoImages addObject:image];
         
-        NSInteger count = self.selectAssets.count;
+        long count = self.selectAssets.count;
         self.makeView.hidden = !count;
         self.makeView.text = [NSString stringWithFormat:@"%ld",(long)count];
         self.doneBtn.enabled = (count > 0);
@@ -300,7 +300,7 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
 }
 
 #pragma mark - setter
--(void)setMaxCount:(NSInteger)maxCount{
+-(void)setMaxCount:(long)maxCount{
     _maxCount = maxCount;
     
     if (!_privateTempMaxCount) {
@@ -340,7 +340,7 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
     
 //    self.selectAssets = [NSMutableArray arrayWithArray:[[NSSet setWithArray:self.selectAssets] allObjects]];
 
-    NSInteger count = self.selectAssets.count;
+    long count = self.selectAssets.count;
     self.makeView.hidden = !count;
     self.makeView.text = [NSString stringWithFormat:@"%ld",(long)count];
     self.doneBtn.enabled = (count > 0);
@@ -353,8 +353,8 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
             asset = deleteAssets;
         }
         
-        NSInteger selectAssetsCurrentPage = -1;
-        for (NSInteger i = 0; i < self.selectAssets.count; i++) {
+        long selectAssetsCurrentPage = -1;
+        for (long i = 0; i < self.selectAssets.count; i++) {
             ZLPhotoAssets *photoAsset = self.selectAssets[i];
             if ([photoAsset isKindOfClass:[UIImage class]]) {
                 continue;
@@ -386,11 +386,11 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
 
 #pragma mark -
 #pragma mark - UICollectionViewDataSource
-- (NSInteger) numberOfSectionsInCollectionView:(UICollectionView *)collectionView{
+- (long) numberOfSectionsInCollectionView:(UICollectionView *)collectionView{
     return 1;
 }
 
-- (NSInteger) collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(NSInteger)section{
+- (long) collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(long)section{
     return self.selectAssets.count;
 }
 
@@ -430,7 +430,7 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
     [self presentViewController:browserVc animated:NO completion:nil];
 }
 
-- (NSInteger)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
+- (long)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
     return self.selectAssets.count;
 }
 
@@ -456,8 +456,8 @@ static NSString *const _identifier = @"toolBarThumbCollectionViewCell";
     
     // 删除选中的照片
     ALAsset *asset = self.selectAssets[indexPath.row];
-    NSInteger currentPage = indexPath.row;
-    for (NSInteger i = 0; i < self.collectionView.dataArray.count; i++) {
+    long currentPage = indexPath.row;
+    for (long i = 0; i < self.collectionView.dataArray.count; i++) {
         ALAsset *photoAsset = self.collectionView.dataArray[i];
         if ([photoAsset isKindOfClass:[ZLPhotoAssets class]] && [asset isKindOfClass:[ZLPhotoAssets class]]) {
                 ZLPhotoAssets *photoAssets = (ZLPhotoAssets *)photoAsset;
